Move test suite to gem (#52)
This pull request moves the test suite into the `gem/` directory and
splits up the big `converter_test.rb` file in multiple smaller files.

This has the benefit that the test suite is being executed in the
context of the gem and not in the context of the Rails app. This change
should also ensure pure Ruby-compatibility now.