Refactor Code Generation
Now that SyntaxTree handles a bunch of formatting for use
we can simplify the code generation quite a bit.

I also use this chance the clean up the code a bit, defining
new helpers and removing unused ones.

The detection of the Nokogiri nodes now also uses Pattern
Matching, which makes is quite more readable.