Enhance `RubyAnalyzer` to detect `instance_methods` and more (#72)
This pull request enhances `Phlexing::RubyAnalyzer` to also detect
`const`, `calls` and `instance_methods`.

Detecting `instance_methods` could to be useful if we want to include
certain modules based on the methods used in the template. For example
we could detect Rails helpers this way and include the appropriate ones
in the generated code.

Detecting `consts` within the template could be useful down the road if
we are dealing with the rendering of other components in the template
and if we want to do some checks around it.

I'm not sure yet what we could do with the `calls`, but I think it might
be useful to have in any case.